fstar kyber usize workaround (#108)

+++ src/kem/kyber/serialize.rs
@@ -51,7 +51,7 @@ pub(super) fn serialize_little_endian<const COMPRESSION_FACTOR: usize, const OUT
         OUT_LEN
     );
 
-    match COMPRESSION_FACTOR {
+    match COMPRESSION_FACTOR as u32 {
         1 => serialize_little_endian_1(re),
         // VECTOR_V_COMPRESSION_FACTOR_768 & VECTOR_V_COMPRESSION_FACTOR_512
         4 => serialize_little_endian_4(re),
@@ -75,7 +75,7 @@ pub(super) fn deserialize_little_endian<const COMPRESSION_FACTOR: usize>(
         (COEFFICIENTS_IN_RING_ELEMENT * COMPRESSION_FACTOR) / 8
     );
 
-    match COMPRESSION_FACTOR {
+    match COMPRESSION_FACTOR as u32 {
         1 => deserialize_little_endian_1(serialized),
         // VECTOR_V_COMPRESSION_FACTOR_768 & VECTOR_V_COMPRESSION_FACTOR_512
         4 => deserialize_little_endian_4(serialized),
